+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/ContentProcessors/AbstractContentParser.php
@@ -170,6 +170,7 @@ abstract class AbstractContentParser implements ContentParserInterface
                 ContentParserOptions::SUPPORT_MARKDOWN_LINKS => true,
                 ContentParserOptions::ADD_CLASSES => true,
                 ContentParserOptions::EMBED_VIDEOS => true,
+                ContentParserOptions::PRETTIFY_CODE => true,
                 ContentParserOptions::TAB_CONTENT => false,
             ],
             $options
@@ -197,6 +198,10 @@ abstract class AbstractContentParser implements ContentParserInterface
         // Append video embeds
         if ($options[ContentParserOptions::EMBED_VIDEOS] ?? null) {
             $htmlContent = $this->embedVideos($htmlContent);
+        }
+        // Prettify code
+        if ($options[ContentParserOptions::PRETTIFY_CODE] ?? null) {
+            $htmlContent = $this->prettifyCode($htmlContent);
         }
         // Convert the <h2> into tabs
         if ($options[ContentParserOptions::TAB_CONTENT] ?? null) {
@@ -372,6 +377,18 @@ abstract class AbstractContentParser implements ContentParserInterface
                     $matches[0]
                     . '<div class="video-responsive-container">' . $videoHTML . '</div>';
             },
+            $htmlContent
+        );
+    }
+
+    /**
+     * Use Google's code-prettify to add styles to <pre><code>
+     */
+    protected function prettifyCode(string $htmlContent): string
+    {
+        return str_replace(
+            '<pre><code class="',
+            '<pre class="prettyprint"><code class="language-',
             $htmlContent
         );
     }