+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/ContentProcessors/ContentParserOptions.php
@@ -11,5 +11,6 @@ class ContentParserOptions
     public const APPEND_PATH_URL_TO_ANCHORS = 'appendPathURLToAnchors';
     public const ADD_CLASSES = 'addClasses';
     public const EMBED_VIDEOS = 'embedVideos';
+    public const PRETTIFY_CODE = 'prettifyCode';
     public const TAB_CONTENT = 'tabContent';
 }